convertTokens remove deprecated permission refs (#291)
* convertTokens remove deprecated permission refs

* optimise table rolling, minor styling changes

* handle old currency for player and in trade

* improved currency handling

+++ src/modules/helper/LootSheetNPC5eHelper.js
@@ -131,7 +131,7 @@ class LootSheetNPC5eHelper {
                             id: tradeItem.dataset.id,
                             data: {
                                 quantity: parseInt(tradeItem.dataset.quantity),
-                                price: parseInt(tradeItem.dataset.price)
+                                price: parseFloat(tradeItem.dataset.price)
                             }
                         },
                     };
